+++ src/main/java/io/github/opendonationassistant/integration/max/model/User.java
@@ -4,4 +4,4 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import io.micronaut.serde.annotation.Serdeable;
 
 @Serdeable
-public record User(@JsonProperty("user_id") Integer userId) {}
+public record User(@JsonProperty("user_id") Long userId) {}
